Reorder tabs
I tried to order them with my best guess of which will be most sought
out by the broader community (i.e., putting documentation earlier).

I also removed the links to the telecon room and the notes from weekly
meetups.  The telecon room changes occasionally, and it's easy to
forget to update it.  It's still in the README file in the main
repository.